requirements.txt
streamlit
psycopg2-binary
torch
open_clip_torch
pillow
openai-whisper
transformers
